ASAA-192 - PR fixes
