--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/Dto.kt
@@ -1,6 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model
-
-/**
- * Represents a Data Transfer Object (DTO).
- */
-interface Dto
