--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/airquality/ColorDto.kt
@@ -1,12 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.airquality
-
-import com.bottlerocketstudios.launchpad.google.utils.data.model.Dto
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ColorDto(
-    val red: Double?,
-    val green: Double?,
-    val blue: Double?,
-    val alpha: Double?
-) : Dto
