--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/place/PlaceAutocompleteResultModel.kt
@@ -1,10 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.place
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PlaceAutocompleteResultModel(
-    val primaryText: String,
-    val secondaryText: String,
-    val placeId: String
-)
