--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/place/PlacePredictionDto.kt
@@ -1,13 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.place
-
-import com.bottlerocketstudios.launchpad.google.utils.data.model.Dto
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PlacePredictionDto(
-    val place: String?,
-    val placeId: String?,
-    val text: TextResponseDto?,
-    val structuredFormat: StructuredFormatDto?,
-    val types: List<String?>
-) : Dto
