--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/place/PlacesAutocompleteRequestDto.kt
@@ -1,9 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.place
-
-import com.bottlerocketstudios.launchpad.google.utils.data.model.Dto
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PlacesAutocompleteRequestDto(
-    val input: String
-) : Dto
