--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/place/StructuredFormatDto.kt
@@ -1,9 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.place
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class StructuredFormatDto(
-    val mainText: TextResponseDto?,
-    val secondaryText: TextResponseDto?
-)
