--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/place/SuggestionDto.kt
@@ -1,8 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.place
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SuggestionDto(
-    val placePrediction: PlacePredictionDto?
-)
