--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/model/place/TextResponseDto.kt
@@ -1,9 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.model.place
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TextResponseDto(
-    val text: String?,
-    val matches: List<MatchDto?>?
-)
