--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/network/service/airquality/AirQualityApiService.kt
@@ -1,8 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.network.service.airquality
-
-import com.bottlerocketstudios.launchpad.google.utils.data.model.airquality.CurrentConditionsResponseDto
-
-interface AirQualityApiService {
-    fun setApiKey(apiKey: String)
-    suspend fun getCurrentConditions(latitude: Double, longitude: Double): CurrentConditionsResponseDto
-}
