--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/network/service/apiKey/ApiKeyService.kt
@@ -1,5 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.network.service.apiKey
-
-interface ApiKeyService {
-    var apiKey: String
-}
