--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/network/serviceImpl/apiKey/ApiKeyServiceImpl.kt
@@ -1,7 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.network.serviceImpl.apiKey
-
-import com.bottlerocketstudios.launchpad.google.utils.data.network.service.apiKey.ApiKeyService
-import org.koin.core.component.KoinComponent
-
-@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-expect class ApiKeyServiceImpl() : ApiKeyService, KoinComponent
