--- kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/data/network/serviceImpl/place/PlacesApiServiceImpl.kt
@@ -1,7 +1,0 @@
-package com.bottlerocketstudios.launchpad.google.utils.data.network.serviceImpl.place
-
-import com.bottlerocketstudios.launchpad.google.utils.data.network.service.places.PlacesApiService
-import io.ktor.client.request.get
-
-@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-expect class PlacesApiServiceImpl() : PlacesApiService
