+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/Dto.kt
@@ -1,0 +1,6 @@
+package com.bottlerocketstudios.launchpad.google.utils.model
+
+/**
+ * Represents a Data Transfer Object (DTO).
+ */
+interface Dto
