+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/LatLngDto.kt
@@ -1,0 +1,10 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import com.bottlerocketstudios.launchpad.google.utils.model.Dto
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LatLngDto(
+    val latitude: Double,
+    val longitude: Double
+) : Dto
