+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/PlaceAutocompleteResultModel.kt
@@ -1,0 +1,10 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PlaceAutocompleteResultModel(
+    val primaryText: String,
+    val secondaryText: String,
+    val placeId: String
+)
