+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/PlaceLocationResponseDto.kt
@@ -1,0 +1,11 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import com.bottlerocketstudios.launchpad.google.utils.model.Dto
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PlaceLocationResponseDto(
+    val id: String,
+    val location: LatLngDto,
+    val displayName: DisplayNameDto
+) : Dto
