+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/PlacePredictionResponseDto.kt
@@ -1,0 +1,9 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import com.bottlerocketstudios.launchpad.google.utils.model.Dto
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PlacePredictionResponseDto(
+    val suggestions: List<SuggestionDto>
+) : Dto
