+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/StructuredFormatDto.kt
@@ -1,0 +1,9 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class StructuredFormatDto(
+    val mainText: TextResponseDto?,
+    val secondaryText: TextResponseDto?
+)
