+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/SuggestionDto.kt
@@ -1,0 +1,8 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SuggestionDto(
+    val placePrediction: PlacePredictionDto?
+)
