+++ kmp-launchpad-google-utils/src/commonMain/kotlin/com/bottlerocketstudios/launchpad/google/utils/model/place/TextResponseDto.kt
@@ -1,0 +1,9 @@
+package com.bottlerocketstudios.launchpad.google.utils.model.place
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TextResponseDto(
+    val text: String?,
+    val matches: List<MatchDto?>?
+)
